auth-gateway/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
kubernetes==31.0.0
python-multipart==0.0.20
boto3
redis>=5.0
supervisor
cryptography>=42.0
pytest>=8.0
pytest-asyncio>=0.23
markdown
prometheus_client>=0.20.0
prometheus_client>=0.20.0
